+++ src/main/kotlin/uk/gov/justice/digital/hmpps/communitypaybackapi/dto/CourseCompletionResolutionDto.kt
@@ -14,7 +14,7 @@ data class CourseCompletionResolutionDto(
 }
 
 data class CourseCompletionCreditTimeDetailsDto(
-  val deliusEventNumber: Long,
+  val deliusEventNumber: Int,
   val appointmentIdToUpdate: Long?,
   val date: LocalDate,
   @field:Min(value = 1)
